Fix for consistent Video URLs
Added a new tool, that will fetch from the Database. It is
pretty generic so maybe able to scale well to other agents.

Essentially, the hypothesis is that the LLM really doesn't like
the really long URLS and when it recalls them from conversational
Fix for consistent Video URLs

Added a new tool, that will fetch from the Database. It is
pretty generic so maybe able to scale well to other agents.

Essentially, the hypothesis is that the LLM really doesn't like
the really long URLS and when it recalls them from conversational
context there is a probability that one of the bytes is slightly different.