Creates a possibly pointless cache of media assets in GCS
Why pointless? Maybe not fully.

In the process I discovered that we were looking up moment_ids in firestore
without a time stamp. This meant we'd pull anyone that matched, and that could be from
3 days ago. Expired. No Good whatsoever.

However, it does give a great fallback.

Also in this commit is improvements to state management. We now after a change to a rundown
(CUEZ) we fetch the data again and put it in state, instead of rely on the Chat Event
context that sits in the native LLM memory. This made significant improvements to recall of
blocks etc the longer the chat went on, without instructing the orchestrator to go fetch
the episode again. Winning. In local testing it is much more rugged, but still not bullet proof.

Finally, the **only** way to actually have this solid is not to pass file uris and Data in the TextPart
where the LLM can munge it, especially if it isn't language and a huge long signed URL.

I have disabled FilePart sending but intend to enable it once I know CUEZ Rundown and EVS will handle
it without braking.