Fix the video URI problem.
Tried not giving the LLM one, and it would Hallucinate one, so
the solution...

To intercept the payload and remove it with something that contains
the asset ID and looks like a video URI. It is short. The LLM gets it correct.

Store it in state (we already store in firestore as a backup).

Look for that uri in requests to other agents, and swap it out with
correct URI.

Solved on orchestrator side. But still failed with some agents, because
their model was also munging the URI, added FilePart for the other agents
to get the video_uri from which doesn't touch the LLM.